+++ programs/dtfs/src/utils/external/folio_program.rs
@@ -68,8 +68,8 @@ impl FolioProgram {
         ctx: Context<UpdateFolio>,
         program_version: Option<Pubkey>,
         program_deployment_slot: Option<u64>,
-        folio_fee: Option<u64>,
-        minting_fee: Option<u64>,
+        folio_fee: Option<u128>,
+        minting_fee: Option<u128>,
         trade_delay: Option<u64>,
         auction_length: Option<u64>,
         fee_recipients_to_add: Vec<FeeRecipient>,
@@ -596,8 +596,8 @@ impl FolioProgram {
         trade_id: u64,
         sell_limit: Range,
         buy_limit: Range,
-        start_price: u64,
-        end_price: u64,
+        start_price: u128,
+        end_price: u128,
         ttl: u64,
     ) -> Result<()> {
         let cpi_program = ctx.accounts.folio_program.to_account_info();
@@ -670,10 +670,10 @@ impl FolioProgram {
 
     pub fn open_trade(
         ctx: Context<OpenTrade>,
-        sell_limit: u64,
-        buy_limit: u64,
-        start_price: u64,
-        end_price: u64,
+        sell_limit: u128,
+        buy_limit: u128,
+        start_price: u128,
+        end_price: u128,
     ) -> Result<()> {
         let cpi_program = ctx.accounts.folio_program.to_account_info();
